Remove dependency to Kernel32.dll
When compiled with various checks, Visual Studio includes code to
forcefully terminate the process if a check fails. It seems that this code
depends on Kernel32.dll, resulting in invalid imports in Dependency
Walker (SUBSYSTEM_EFI binaries should not depend on SUBSYSTEM_WIN32 DLLs),
as well in validation errors when Secure Boot is involved.

I doubt these checks would result in a clean termination of the EFI
program anyway :-)